+++ app/routes/acquisitions.js
@@ -10,79 +10,54 @@ let assert = require('assert');
 
 function acquisitions(req, res, db, args) {
 
-	let sport = args.sport;
-	let year = args.year;
-	let owner_number = args.owner_number;
-	var completed_sport_season = args.completed_sport_season;
-
-	if (sport === 'football') {
-		acquire_process();
-	}
-	else {
-		// scrape PR once for basketball and baseball
-		var pr_scrape = require('../modules/' + sport + '_acquisitions_pr.js')(req, res, db, sport, year, function(err, call2) {
-			console.log("PR scrape done");
-
-			acquire_process();
-		})
-	}
-
+// define functions for module
 var acquire_process = function() {
 
-	// if year is greater than what's been completed, scrape
-	if (year > completed_sport_season) {
+	if (owner_number === 'all') {
 
-		if (owner_number === 'all') {
+		// scrape draft once
+		var draft = require('../modules/draft.js')(req, res, db, sport, year, function(err, call4) {
+			console.log("drafted players done");
 
-			// scrape draft once
-			var draft = require('../modules/draft.js')(req, res, db, sport, year, function(err, call4) {
-				console.log("drafted players done");
+			// set array of numbers
+			var owner_list = ["1", "2", "3", "4", "5", "6", "7", "8", "9", "10"]
 
-				// set array of numbers
-				var owner_list = ["1", "2", "3", "4", "5", "6", "7", "8", "9", "10"]
-
-				// call function that synchronously loops through acquistions for each team
-				all_acquisitions(0, owner_list);
-				
-			}) // end of draft scrape
-		}
-
-		// if owner is a specific owner
-		else {
-			// send to script that adds drafted players
-			var draft = require('../modules/draft.js')(req, res, db, sport, year, function(err, call2) {
-				console.log("drafted players done");			
+			// call function that synchronously loops through acquistions for each team
+			all_acquisitions(0, owner_list);
 			
-				var active_stats = require('../modules/' + sport + '_active_stats.js')(req, res, db, sport, year, owner_number, function(err, call){
-					console.log("active stats scrape done");
-
-					var add = require('../modules/' + sport + '_add.js')(req, res, db, sport, year, owner_number, function(err, call3) {
-						console.log("added players done");
-
-						var aux = require('../modules/' + sport + '_acquisitions_aux.js')(req, res, db, sport, year, owner_number, function(err, call5) {
-							console.log("aux stats added");
-
-							var options = {
-								args: [sport, year, owner_number]
-							}
-
-							// python script that determines acqusition value
-							pyshell.run(sport + '_acquisitions.py', options, function(err) {
-								console.log("acquisition python script complete");
-								var acquisitions_display = require('../modules/acquisitions_display.js')(req, res, db, sport, year, owner_number)			
-							}) // end of python script
-
-						}) // end of aux
-					}) // end of football draft script
-				}) // end of football add script
-			})	// end of footbabll draft
-		} // end of else for specific owner
+		}) // end of draft scrape
 	}
-	// if don't need to scrape, just pull, sort and display
-	else {
-		var acquisitions_display = require('../modules/acquisitions_display.js')(req, res, db, sport, year, owner_number)
 
-	} // end of if don't need to scrape, just dispaly
+	// if owner is a specific owner
+	else {
+		// send to script that adds drafted players
+		var draft = require('../modules/draft.js')(req, res, db, sport, year, function(err, call2) {
+			console.log("drafted players done");			
+		
+			var active_stats = require('../modules/' + sport + '_active_stats.js')(req, res, db, sport, year, owner_number, function(err, call){
+				console.log("active stats scrape done");
+
+				var add = require('../modules/' + sport + '_add.js')(req, res, db, sport, year, owner_number, function(err, call3) {
+					console.log("added players done");
+
+					var aux = require('../modules/' + sport + '_acquisitions_aux.js')(req, res, db, sport, year, owner_number, function(err, call5) {
+						console.log("aux stats added");
+
+						var options = {
+							args: [sport, year, owner_number]
+						}
+
+						// python script that determines acqusition value
+						pyshell.run(sport + '_acquisitions.py', options, function(err) {
+							console.log("acquisition python script complete");
+							var acquisitions_display = require('../modules/acquisitions_display.js')(req, res, db, sport, year, owner_number)			
+						}) // end of python script
+
+					}) // end of aux
+				}) // end of football draft script
+			}) // end of football add script
+		})	// end of footbabll draft
+	} // end of else for specific owner
 } // end of acquire_process function
 
 // function that synchronously updates acquisitions for each team
@@ -124,6 +99,31 @@ var all_acquisitions = function(x, owner_list) {
 		var acquisitions_display = require('../modules/acquisitions_display.js')(req, res, db, sport, year, owner_number)			
 	}
 } // end of all_football_acquisitions function
+
+// execute script //
+	let sport = args.sport;
+	let year = args.year;
+	let owner_number = args.owner_number;
+	var completed_sport_season = args.completed_sport_season;
+
+	if (year > completed_sport_season) {
+
+		if (sport === 'football') {
+			acquire_process();
+		}
+		else {
+			// scrape PR once for basketball and baseball
+			var pr_scrape = require('../modules/' + sport + '_acquisitions_pr.js')(req, res, db, sport, year, function(err, call2) {
+				console.log("PR scrape done");
+
+				acquire_process();
+			})
+		}
+	}
+
+	else {
+		var acquisitions_display = require('../modules/acquisitions_display.js')(req, res, db, sport, year, owner_number);
+	}
 
 } // end of acquisitions module
