+++ app/routes/routes.js
@@ -421,9 +421,7 @@ router.get('/origin_home_page', function(req, res) {
 // route to origin analysis
 router.get('/:sport/origin/:year', function(req, res) {
 
-	// set variables from request url
 	var sport = req.params.sport;
-	var year = req.params.year;
 
 	// set completed season for check if in season or not depending on sport
 	if (sport === 'football') {
@@ -436,246 +434,14 @@ router.get('/:sport/origin/:year', function(req, res) {
 		completed_sport_season = completed_baseball_season;
 	}
 
-	if (year > completed_sport_season) {
-		
-		if (sport == "football") {
+	var input = {
+		sport: req.params.sport,
+		year: req.params.year,
+		completed_sport_season: completed_sport_season
+	};
 
-			// send to script that adds drafted players
-			var football_draft = require('./draft.js')(req, res, db, sport, year, function(err, call2) {
-				console.log("drafted players done");
-
-				// set array of numbers
-				var owner_list = ["1", "2", "3", "4", "5", "6", "7", "8", "9", "10"]				
-
-				// call function that synchronously loops through acquistions for each team
-				all_football_origin(0, owner_list);		
-
-			})	// end of footbabll draft
-
-	// function that synchronously updates acquisitions for each team
-	var all_football_origin = function(x, owner_list) {
-
-		if (x < owner_list.length) {
-
-			owner_number = owner_list[x];
-			console.log("owner number", owner_number);
-
-			var active_stats = require('./football_active_stats.js')(req, res, db, sport, year, owner_number, function(err, call){
-				console.log("active stats scrape done");
-
-				var football_add = require('./football_add.js')(req, res, db, sport, year, owner_number, function(err, call3) {
-					console.log("added players done");
-
-					var football_aux = require('./football_acquisitions_aux.js')(req, res, db, sport, year, owner_number, function(err, call5) {
-						console.log("aux stats added");
-
-						var options = {
-							args: [sport, year, owner_number]
-						}
-
-						pyshell.run('football_origin.py', options, function(err) {
-							console.log("origin python script complete");
-
-							all_football_origin(x + 1, owner_list);
-						
-						}) // end of pyshell 
-					}) // end of football auxiliary
-				}) // end of football add 
-			}) // end of active stats
-		}
-		// if done with all the owners in owners_list
-		else {
-			// reset owner number (after it has gone through loop) to all for all display
-			db.collection(sport + "_origin_" + year).find({}, {"_id": 0}).sort({"total_points": -1}).toArray(function(e, docs) {
-				origin_standings = docs;
-
-				console.log("displaying origin standings...");
-				res.render('origin', {
-					sport: sport,
-					year: year,
-					origin: origin_standings
-				})
-			})		
-		}
-
-	} // end of all_football_acquisitions function			
-		
-		} // end of if football
-
-		else if (sport == 'basketball') {
-
-			// scrape PR once
-			var pr_scrape = require('./basketball_acquisitions_pr.js')(req, res, db, sport, year, function(err, call2) {
-				console.log("PR scrape done");
-
-				// scrape draft once
-				var basketball_draft = require('./draft.js')(req, res, db, sport, year, function(err, call4) {
-					console.log("drafted players done");
-
-					// set array of numbers
-					var owner_list = ["1", "2", "3", "4", "5", "6", "7", "8", "9", "10"]	
-
-					// call function that synchronously loops through acquistions for each team
-					all_basketball_origin(0, owner_list);
-					
-				}) // end of draft scrape
-			}) // end of pr scrape	
-
-
-	// function that synchronously updates acquisitions for each team
-	var all_basketball_origin = function(x, owner_list) {
-
-		if (x < owner_list.length) {
-
-			owner_number = owner_list[x];
-			console.log("owner number", owner_number);
-
-			var active_stats = require('./basketball_active_stats.js')(req, res, db, sport, year, owner_number, function(err, call){
-				console.log("active stats scrape done");
-
-				var basketball_add = require('./basketball_add.js')(req, res, db, sport, year, owner_number, function(err, call3) {
-					console.log("added players done");
-
-					var basketball_aux = require('./basketball_acquisitions_aux.js')(req, res, db, sport, year, owner_number, function(err, call5) {
-						console.log("aux stats added");
-
-						var options = {
-							args: [sport, year, owner_number]
-						}
-
-						pyshell.run('basketball_origin.py', options, function(err) {
-							console.log("origin python script complete");
-
-							all_basketball_origin(x + 1, owner_list);
-						
-						}) // end of pyshell 
-
-					}) // end of basketball auxiliary
-				}) // end of basketball add 
-			}) // end of active stats
-		}
-		// if done with all the owners in owners_list
-		else {
-
-			// reset owner number (after it has gone through loop) to all for all display
-			db.collection(sport + "_origin_" + year).find({}, {"_id": 0}).sort({"total_pr": -1}).toArray(function(e, docs) {
-				origin_standings = docs;
-				console.log("displaying origin standings...");
-				res.render('origin', {
-					sport: sport,
-					year: year,
-					origin: origin_standings
-				})
-			})					
-		}
-
-	} // end of all_basketball_acquisitions function
-		
-		} // end of if basketball
-
-		else if (sport == 'baseball') {
-
-			// scrape draft once
-			var baseball_draft = require('./draft.js')(req, res, db, sport, year, function(err, call4) {
-				console.log("drafted players done");		
-
-				// scrape PRs once
-				var pr_scrape = require('./baseball_acquisitions_pr.js')(req, res, db, sport, year, function(err, call2) {
-					console.log("pr srape done");	
-
-					// set array of numbers
-					var owner_list = ["1", "2", "3", "4", "5", "6", "7", "8", "9", "10"]
-
-					// call function that synchronously loops through acquistions for each team
-					all_baseball_origin(0, owner_list);
-					
-				}) // end of draft scrape
-			}) // end of pr scrape	
-
-	// function that synchronously updates acquisitions for each team
-	var all_baseball_origin = function(x, owner_list) {
-
-		// loop through owner list
-		if (x < owner_list.length) {
-
-			// set owner number
-			owner_number = owner_list[x];
-			console.log("owner number", owner_number);
-
-			// scrape active stats
-			var active_stats = require('./baseball_active_stats.js')(req, res, db, sport, year, owner_number, function(err, call1) {
-				console.log("active stats scrape done");
-
-				// scrape additions
-				var baseball_add = require('./baseball_add.js')(req, res, db, sport, year, owner_number, function(err, call3) {
-					console.log("added players done");
-
-					// add aux stats in
-					var baseball_aux = require('./baseball_acquisitions_aux.js')(req, res, db, sport, year, owner_number, function(err, call5) { 
-						console.log("aux stats done");
-
-						var options = {
-							args: [sport, year, owner_number]
-						}
-
-						// create acquisition values
-						pyshell.run('baseball_origin.py', options, function(err) {
-							console.log("origin python script done");
-							all_baseball_origin(x + 1, owner_list);
-						}) // end of pyshell						
-
-					}) // end of basketball auxiliary
-				}) // end of basketball add 
-			}) // end of active stats
-		}
-		// if done with all the owners in owners_list
-		else {
-			// reset owner number (after it has gone through loop) to all for all display
-			db.collection(sport + "_origin_" + year).find({}, {"_id": 0}).sort({"total_pr": -1}).toArray(function(e, docs) {
-				origin_standings = docs;
-				console.log("displaying origin standings...");
-				res.render('origin', {
-					sport: sport,
-					year: year,
-					origin: origin_standings
-				})
-			})		
-		}
-	} // end of all_basall_acquisitions function
-
-		} // end of if baseball		
-	
-	} // end of if need to scrape
-
-	// if year is in past and just display
-	else {
-
-		if (sport == "football") {
-					// reset owner number (after it has gone through loop) to all for all display
-			db.collection(sport + "_origin_" + year).find({}, {"_id": 0}).sort({"total_points": -1}).toArray(function(e, docs) {
-				origin_standings = docs;
-				console.log("displaying origin standings...");
-				res.render('origin', {
-					sport: sport,
-					year: year,
-					origin: origin_standings
-				})
-			})			
-		}
-
-		else {
-			// reset owner number (after it has gone through loop) to all for all display
-			db.collection(sport + "_origin_" + year).find({}, {"_id": 0}).sort({"total_pr": -1}).toArray(function(e, docs) {
-				origin_standings = docs;
-				console.log("displaying origin standings...");
-				res.render('origin', {
-					sport: sport,
-					year: year,
-					origin: origin_standings
-				})
-			})	
-		}		
-	} // end of if just display
+	let origin = require('./origin.js').origin;
+	const send = origin(req, res, db, input);
 
 }); // end of origin route
